Add form generator
